fix admin

+++ app/src/main/java/com/app_computer_ecom/dack/screen/admin/BannerScreen.kt
@@ -6,11 +6,17 @@ import android.util.Log
 import androidx.activity.compose.BackHandler
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -21,6 +27,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
+import com.app_computer_ecom.dack.LoadingScreen
 import com.app_computer_ecom.dack.components.ImagePreviewItem
 import com.app_computer_ecom.dack.components.admin.HeaderViewMenu
 import com.app_computer_ecom.dack.model.BannerModel
@@ -43,6 +50,7 @@ fun BannerScreen(navController: NavHostController, modifier: Modifier = Modifier
     var banners by remember {
         mutableStateOf(emptyList<BannerModel>())
     }
+    var isLoading by remember { mutableStateOf(true) }
 
     BackHandler(enabled = true) {
         navController.navigate("admin/01") {
@@ -55,6 +63,7 @@ fun BannerScreen(navController: NavHostController, modifier: Modifier = Modifier
 
     LaunchedEffect(Unit) {
         banners = bannerRepository.getBanners()
+        isLoading = false
     }
 
     val launcher = rememberLauncherForActivityResult(
@@ -100,34 +109,50 @@ fun BannerScreen(navController: NavHostController, modifier: Modifier = Modifier
             isUploading = isUploading.value
         )
 
-        LazyColumn {
-            items(banners.size) { index ->
-                ImagePreviewItem(
-                    uri = banners[index].imageUrl,
-                    onClick = {
-                        CoroutineScope(Dispatchers.IO).launch {
-                            try {
-                                ImageCloudinary.deleteImage(banners[index].imageUrl)
-                                bannerRepository.deleteBanner(banners[index])
-                                banners = bannerRepository.getBanners()
-                            } catch (e: Exception) {
-                                Log.e("BannerScreen", "Delete failed: ${e.message}")
-                            }
-                        }
-                    },
-                    onClickShowHidden = {
-                        CoroutineScope(Dispatchers.IO).launch {
-                            try {
-                                bannerRepository.updateBanner(banners[index].copy(enable = !banners[index].enable))
-                                banners = bannerRepository.getBanners()
-                            } catch (e: Exception) {
+        if (isLoading) {
+            LoadingScreen()
+        } else {
+            LazyColumn(
+                modifier = Modifier.padding(horizontal = 16.dp),
+                verticalArrangement = Arrangement.spacedBy(8.dp),
+            ) {
+                items(banners.size) { index ->
+                    Card(
+                        shape = RoundedCornerShape(12.dp),
+                        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface),
+                        elevation = CardDefaults.cardElevation(8.dp),
+                        modifier = Modifier
+                            .fillMaxWidth()
+                    ) {
+                        ImagePreviewItem(
+                            uri = banners[index].imageUrl,
+                            onClick = {
+                                CoroutineScope(Dispatchers.IO).launch {
+                                    try {
+                                        ImageCloudinary.deleteImage(banners[index].imageUrl)
+                                        bannerRepository.deleteBanner(banners[index])
+                                        banners = bannerRepository.getBanners()
+                                    } catch (e: Exception) {
+                                        Log.e("BannerScreen", "Delete failed: ${e.message}")
+                                    }
+                                }
+                            },
+                            onClickShowHidden = {
+                                CoroutineScope(Dispatchers.IO).launch {
+                                    try {
+                                        bannerRepository.updateBanner(banners[index].copy(enable = !banners[index].enable))
+                                        banners = bannerRepository.getBanners()
+                                    } catch (e: Exception) {
 
-                            }
-                        }
-                    },
-                    isEnable = banners[index].enable,
-                    modifier = Modifier.fillMaxWidth().height(120.dp)
-                )
+                                    }
+                                }
+                            },
+                            isEnable = banners[index].enable,
+                            modifier = Modifier.fillMaxWidth().height(120.dp)
+                        )
+                    }
+
+                }
             }
         }
